solidfire-libs eclass: don't version binaries by default
Older versions of solidfire-libs used to append "-solidfire-${PV}" to
the names of the binaries, to allow multiple versions of the binaries
to be installed side by side.

We now install all solidfire-libs ebuilds into versioned directories,
so this is no longer needed.

Individual ebuilds can set PS to get the old behavior.